ouygbiugh

+++ my_average.c
@@ -1,6 +1,6 @@
 float my_average(int *grades, int count) {
-    int total = 0;
-    float average;
+    long long int total = 0;
+    double average;
 
     for (int i = 0; i<count; i++) {
         total += grades[i];
